+++ src/smart_letters/config.py
@@ -3,11 +3,13 @@ import json
 from functools import wraps
 from pathlib import Path
 from typing import Annotated, Any
+import os
 
 import snick
 import typer
+from inflection import dasherize
 from loguru import logger
-from pydantic import AfterValidator, BaseModel, ValidationError
+from pydantic import AfterValidator, BaseModel, ValidationError, Field
 
 from smart_letters.exceptions import Abort, handle_abort
 from smart_letters.cache import CACHE_DIR, init_cache
@@ -27,6 +29,16 @@ def file_exists(value: Path | None) -> Path | None:
     return value
 
 
+def has_editor(value: str | None) -> str:
+    if value is not None:
+        return value
+
+    try:
+        return os.environ["EDITOR"]
+    except Exception:
+        raise ValueError("Couldn't load editor from environment. Please set it explicitly")
+
+
 class Settings(BaseModel):
     openai_api_key: str
     resume_path: Annotated[Path, AfterValidator(file_exists)]
@@ -34,6 +46,22 @@ class Settings(BaseModel):
     filename_prefix: str = "cover-letter"
     heading_path: Annotated[Path | None, AfterValidator(file_exists)] = None
     sig_path: Annotated[Path | None, AfterValidator(file_exists)] = None
+    output_directory: Annotated[Path | None, AfterValidator(file_exists)] = None
+    dev_prompt_path: Annotated[Path | None, AfterValidator(file_exists)] = None
+    user_prompt_template_path: Annotated[Path | None, AfterValidator(file_exists)] = None
+    editor_command: Annotated[str | None, AfterValidator(has_editor)] = None
+
+    invalid_warning: Annotated[
+        str | None,
+        Field(
+            exclude=True,
+            description="""
+            An optional warning that can be included when the model is invalid.
+
+            Used when we use the `attach_settings` decorator with `validate=False`.
+        """,
+        ),
+    ] = None
 
 
 @contextmanager
@@ -54,51 +82,71 @@ def handle_config_error():
         )
 
 
-def init_settings(**settings_values) -> Settings:
+def init_settings(validate: bool = True, **settings_values) -> Settings:
     with handle_config_error():
         logger.debug("Validating settings")
-        return Settings(**settings_values)
+        try:
+            return Settings(**settings_values)
+        except ValidationError as err:
+            if validate:
+                raise
+            settings = Settings.model_construct(**settings_values)
+            settings.invalid_warning = str(err)
+            return settings
 
 
 def update_settings(settings: Settings, **settings_values) -> Settings:
     with handle_config_error():
         logger.debug("Validating settings")
-        return settings.model_copy(update=settings_values)
+        settings_dict = settings.model_dump(exclude_unset=True)
+        settings_dict.update(**settings_values)
+        return Settings(**settings_dict)
 
 
 def unset_settings(settings: Settings, *unset_keys) -> Settings:
     with handle_config_error():
         logger.debug("Unsetting settings")
-        return Settings(
-            **{
-                k: v
-                for (k, v) in settings.model_dump(exclude_unset=True).items()
-                if k not in unset_keys
-            }
-        )
+        return Settings(**{k: v for (k, v) in settings.model_dump(exclude_unset=True).items() if k not in unset_keys})
 
 
-def attach_settings(func):
-    @wraps(func)
-    def wrapper(ctx: typer.Context, *args, **kwargs):
-        try:
-            logger.debug(f"Loading settings from {settings_path}")
-            settings_values = json.loads(settings_path.read_text())
-        except FileNotFoundError:
-            raise Abort(
-                f"""
-                No settings file found at {settings_path}!
+def attach_settings(original_function=None, *, validate=True):
+    """
+    Attach the settings to the CLI context.
 
-                Run the set-config sub-command first to establish your settings.
-                """,
-                subject="Settings file missing!",
-                log_message="Settings file missing!",
-            )
-        logger.debug("Binding settings to CLI context")
-        ctx.obj.settings = init_settings(**settings_values)
-        return func(ctx, *args, **kwargs)
+    Optionally, skip validation of the settings. This is useful in case the config
+    file being loaded is not valid, but we still want to use the settings. Then, we
+    can update the settings with correct values.
 
-    return wrapper
+    Uses recipe for decorator with optional arguments from:
+    https://stackoverflow.com/a/24617244/642511
+    """
+
+    def _decorate(func):
+        @wraps(func)
+        def wrapper(ctx: typer.Context, *args, **kwargs):
+            try:
+                logger.debug(f"Loading settings from {settings_path}")
+                settings_values = json.loads(settings_path.read_text())
+            except FileNotFoundError:
+                raise Abort(
+                    f"""
+                    No settings file found at {settings_path}!
+
+                    Run the set-config sub-command first to establish your settings.
+                    """,
+                    subject="Settings file missing!",
+                    log_message="Settings file missing!",
+                )
+            logger.debug("Binding settings to CLI context")
+            ctx.obj.settings = init_settings(validate=validate, **settings_values)
+            return func(ctx, *args, **kwargs)
+
+        return wrapper
+
+    if original_function:
+        return _decorate(original_function)
+    else:
+        return _decorate
 
 
 def dump_settings(settings: Settings):
@@ -112,28 +160,56 @@ def clear_settings():
     settings_path.unlink(missing_ok=True)
 
 
-cli = typer.Typer()
+cli = typer.Typer(help="Configure the app, change settings, or view how it's currently configured")
 
 
 @cli.command()
 @handle_abort
 @init_cache
 def bind(
-    openai_api_key: Annotated[
-        str, typer.Option(help="The API key needed to access OpenAI")
-    ],
+    openai_api_key: Annotated[str, typer.Option(help="The API key needed to access OpenAI")],
     resume_path: Annotated[Path, typer.Option(help="The path to your resume")],
     candidate_name: Annotated[
-        str, typer.Option(help="The name of the candidate to use in the closing")
+        str,
+        typer.Option(help="The name of the candidate to use in the closing"),
     ],
     filename_prefix: Annotated[
-        str, typer.Option(help="The filename prefix to use for your cover letter")
+        str,
+        typer.Option(help="The filename prefix to use for your cover letter"),
     ] = "cover-letter",
     sig_path: Annotated[
-        Path | None, typer.Option(help="The path to your signature")
+        Path | None,
+        typer.Option(help="The path to your signature"),
     ] = None,
     heading_path: Annotated[
-        Path | None, typer.Option(help="The path to your markdown heading")
+        Path | None,
+        typer.Option(help="The path to your markdown heading"),
+    ] = None,
+    output_directory: Annotated[
+        Path | None,
+        typer.Option(
+            help="""
+                The directory where cover letters should be saved.
+                If unset, letters will be saved in the current directory
+            """
+        ),
+    ] = None,
+    dev_prompt_path: Annotated[
+        Path | None,
+        typer.Option(help="An optional path to the developer prompt for letter generation"),
+    ] = None,
+    user_prompt_template_path: Annotated[
+        Path | None,
+        typer.Option(help="An optional path to the user prompt template for letter generation"),
+    ] = None,
+    editor_command: Annotated[
+        str | None,
+        typer.Option(
+            help="""
+                Provide a system command to use for opening files in an editor.
+                If not provided, will attempt to load system default editor (from env var $EDITOR).
+            """
+        ),
     ] = None,
 ):
     """
@@ -147,6 +223,10 @@ def bind(
         filename_prefix=filename_prefix,
         sig_path=sig_path,
         heading_path=heading_path,
+        output_directory=output_directory,
+        dev_prompt_path=dev_prompt_path,
+        user_prompt_path=user_prompt_template_path,
+        editor_command=editor_command,
     )
     dump_settings(settings)
 
@@ -154,47 +234,50 @@ def bind(
 @cli.command()
 @handle_abort
 @init_cache
-@attach_settings
+@attach_settings(validate=False)
 def update(
     ctx: typer.Context,
-    openai_api_key: Annotated[
-        str | None, typer.Option(help="The API key needed to access OpenAI")
-    ] = None,
-    resume_path: Annotated[
-        Path | None, typer.Option(help="The path to your resume")
-    ] = None,
-    candidate_name: Annotated[
-        str | None, typer.Option(help="The name of the candidate to use in the closing")
-    ] = None,
+    openai_api_key: Annotated[str | None, typer.Option(help="The API key needed to access OpenAI")] = None,
+    resume_path: Annotated[Path | None, typer.Option(help="The path to your resume")] = None,
+    candidate_name: Annotated[str | None, typer.Option(help="The name of the candidate to use in the closing")] = None,
     filename_prefix: Annotated[
         str | None,
         typer.Option(help="The filename prefix to use for your cover letter"),
     ] = None,
-    sig_path: Annotated[
-        Path | None, typer.Option(help="The path to your signature")
+    sig_path: Annotated[Path | None, typer.Option(help="The path to your signature")] = None,
+    heading_path: Annotated[Path | None, typer.Option(help="The path to your markdown heading")] = None,
+    output_directory: Annotated[
+        Path | None,
+        typer.Option(
+            help="""
+                The directory where cover letters should be saved.
+                If unset, letters will be saved in the current directory
+            """
+        ),
     ] = None,
-    heading_path: Annotated[
-        Path | None, typer.Option(help="The path to your markdown heading")
+    dev_prompt_path: Annotated[
+        Path | None,
+        typer.Option(help="An optional path to the developer prompt for letter generation"),
+    ] = None,
+    user_prompt_template_path: Annotated[
+        Path | None,
+        typer.Option(help="An optional path to the user prompt template for letter generation"),
+    ] = None,
+    editor_command: Annotated[
+        str | None,
+        typer.Option(
+            help="""
+                Provide a system command to use for opening files in an editor.
+                If not provided, will attempt to load system default editor (from env var $EDITOR).
+            """
+        ),
     ] = None,
 ):
     """
-    Bind the configuration to the app.
+    Update one or more configuration settings that are bound to the app.
     """
     logger.debug(f"Updating settings with {locals()}")
-    kwargs: dict[str, Any] = {}
-    if openai_api_key is not None:
-        kwargs["openai_api_key"] = openai_api_key
-    if resume_path is not None:
-        kwargs["resume_path"] = resume_path
-    if candidate_name is not None:
-        kwargs["candidate_name"] = candidate_name
-    if filename_prefix is not None:
-        kwargs["filename_prefix"] = filename_prefix
-    if sig_path is not None:
-        kwargs["sig_path"] = sig_path
-    if heading_path is not None:
-        kwargs["heading_path"] = heading_path
-
+    kwargs: dict[str, Any] = {k: v for (k, v) in locals().items() if v is not None}
     settings = update_settings(ctx.obj.settings, **kwargs)
     dump_settings(settings)
 
@@ -202,26 +285,44 @@ def update(
 @cli.command()
 @handle_abort
 @init_cache
-@attach_settings
+@attach_settings(validate=False)
 def unset(
     ctx: typer.Context,
-    openai_api_key: Annotated[
-        bool, typer.Option(help="The API key needed to access OpenAI")
-    ] = False,
+    openai_api_key: Annotated[bool, typer.Option(help="The API key needed to access OpenAI")] = False,
     resume_path: Annotated[bool, typer.Option(help="The path to your resume")] = False,
-    candidate_name: Annotated[
-        bool, typer.Option(help="The name of the candidate to use in the closing")
-    ] = False,
-    filename_prefix: Annotated[
-        bool, typer.Option(help="The filename prefix to use for your cover letter")
-    ] = False,
+    candidate_name: Annotated[bool, typer.Option(help="The name of the candidate to use in the closing")] = False,
+    filename_prefix: Annotated[bool, typer.Option(help="The filename prefix to use for your cover letter")] = False,
     sig_path: Annotated[bool, typer.Option(help="The path to your signature")] = False,
-    heading_path: Annotated[
-        bool, typer.Option(help="The path to your markdown heading")
+    heading_path: Annotated[bool, typer.Option(help="The path to your markdown heading")] = False,
+    output_directory: Annotated[
+        bool,
+        typer.Option(
+            help="""
+                The directory where cover letters should be saved.
+                If unset, letters will be saved in the current directory
+            """
+        ),
+    ] = False,
+    dev_prompt_path: Annotated[
+        bool,
+        typer.Option(help="An optional path to the developer prompt for letter generation"),
+    ] = False,
+    user_prompt_template_path: Annotated[
+        bool,
+        typer.Option(help="An optional path to the user prompt template for letter generation"),
+    ] = False,
+    editor_command: Annotated[
+        bool,
+        typer.Option(
+            help="""
+                Provide a system command to use for opening files in an editor.
+                If not provided, will attempt to load system default editor (from env var $EDITOR).
+            """
+        ),
     ] = False,
 ):
     """
-    Bind the configuration to the app.
+    Remove a configuration setting that was previously bound to the app.
     """
     logger.debug(f"Updating settings with {locals()}")
     keys = [k for k in locals() if locals()[k]]
@@ -232,16 +333,20 @@ def unset(
 @cli.command()
 @handle_abort
 @init_cache
-@attach_settings
+@attach_settings(validate=False)
 def show(ctx: typer.Context):
     """
     Show the config that is currently bound to the app.
     """
     parts = []
     for field_name, field_value in ctx.obj.settings:
-        parts.append((field_name, field_value))
+        if field_name == "invalid_warning":
+            continue
+        parts.append((dasherize(field_name), field_value))
     max_field_len = max(len(field_name) for field_name, _ in parts)
     message = "\n".join(f"[bold]{k:<{max_field_len}}[/bold] -> {v}" for k, v in parts)
+    if ctx.obj.settings.invalid_warning:
+        message += f"\n\n[red]Configuration is invalid: {ctx.obj.settings.invalid_warning}[/red]"
     terminal_message(message, subject="Current Configuration")