+++ src/smart_letters/generate.py
@@ -1,210 +1,201 @@
+from dataclasses import dataclass
 import json
-import os
 import subprocess
-import tempfile
-from importlib import resources
 from pathlib import Path
-from typing import Annotated, Any
+from typing import Annotated
 
 import httpx
 import typer
+import pyperclip
 from inflection import parameterize
 from loguru import logger
-from markdown import markdown
 from openai import OpenAI
-from weasyprint import HTML
 
+from smart_letters.cache import BACKUP_DIR, init_cache
 from smart_letters.config import attach_settings
 from smart_letters.exceptions import handle_abort
 from smart_letters.format import terminal_message, simple_message
+from smart_letters.prompt import get_prompts, build_prompts
+from smart_letters.utilities import asset_path
+from smart_letters.render import render_letter
+from smart_letters.schemas import LetterConfig, PromptConfig, RenderConfig
+
+
+@dataclass
+class Reprompt:
+    old_letter: str
+    user_feedback: str
+
+
+def pull_posting(letter_config: LetterConfig) -> str:
+    logger.debug(f"Pulling posting from {letter_config.posting_url}")
+    response = httpx.get(letter_config.posting_url)
+    response.raise_for_status()
+    return response.text
+
+
+def pull_resume(letter_config: LetterConfig) -> str:
+    logger.debug(f"Loading resume from {letter_config.resume_path}")
+    return letter_config.resume_path.read_text()
+
+
+def pull_example(letter_config: LetterConfig) -> str | None:
+    if letter_config.example_path:
+        logger.debug(f"Loading example text from {letter_config.example_path}")
+        return letter_config.example_path.read_text()
+    return None
+
+
+def pull_heading(letter_config: LetterConfig) -> str | None:
+    if letter_config.heading_path:
+        logger.debug(f"Loading heading text from {letter_config.heading_path}")
+        return letter_config.heading_path.read_text()
+    return None
 
 
 def generate_letter(
-    api_key: str,
+    prompt_config: PromptConfig,
+    letter_config: LetterConfig,
+    reprompts: list[Reprompt],
     posting_text: str,
     resume_text: str,
     example_text: str | None = None,
-    reprompts: list[tuple[str, str]] | None = None,
 ) -> str:
-    client = OpenAI(api_key=api_key)
+    logger.debug("Generating letter using OpenAI")
+    client = OpenAI(api_key=letter_config.openai_api_key)
 
-    messages = []
-
-    messages.append(
-        dict(
-            role="developer",
-            content="""
-                You are a professional resume writer and career coach. Your task is to
-                generate a cover letter for a job application based on the provided job
-                posting and resume. You will be given a job posting and a resume, and
-                you need to create a tailored cover letter that highlights the
-                candidate's qualifications and suitability for the position.
-            """,
-        ),
+    prompts = get_prompts(prompt_config)
+    messages = build_prompts(
+        prompts,
+        posting_text=posting_text,
+        resume_text=resume_text,
+        example_text=example_text,
     )
 
-    messages.append(
-        dict(
-            role="developer",
-            content=f"""
-                This is the job posting. It is html scraped from a job board. It
-                contains a lot of html markup, but you should ignore that and focus on
-                the text content.
+    for reprompt in reprompts:
+        messages.append(dict(role="assistant", content=reprompt.old_letter))
+        messages.append(dict(role="user", content=reprompt.user_feedback))
 
-                {posting_text}
-            """,
-        ),
-    )
-
-    messages.append(
-        dict(
-            role="developer",
-            content=f"""
-                Use this resume as you write the letter. It highlights the skills and
-                experience of the candidate.
-
-                {resume_text}
-            """,
-        ),
-    )
-
-    if example_text:
-        messages.append(
-            dict(
-                role="developer",
-                content=f"""
-                    You should use this cover letter as reference for what I would like
-                    the resulting cover letter to look like. It is a markdown document.
-
-                    {example_text}
-                """,
-            ),
-        )
-
-    messages.append(
-        dict(
-            role="developer",
-            content="""
-                The cover letter should not sound like it was written by a machine. It
-                should be direct, personal, and energetic. It should be professional,
-                but informal in an almost daring way.
-
-                Avoid sentences of the form, "With X, I will Y". Do not use any of the
-                following words:
-
-                - challenge
-                - opportunity
-                - leverage
-                - passionate
-
-                The letter should highlight the candidate's leadership, technical
-                excellence, and ambition.
-
-                Do not include a salutation or closing. Only include the body of the
-                letter. Don't include any other text. Just print the body of the letter
-                itself.
-
-                Again, do not provide any friendly text before the letter. Do not
-                explain what you are doing. Do not say "Here is your letter". Just print
-                the damn letter.
-            """,
-        )
-    )
-
-    messages.append(dict(role="user", content="Please write me a nice letter!"))
-
-    if reprompts is not None:
-        for old_letter, user_feedback in reprompts:
-            messages.append(dict(role="assistant", content=old_letter))
-            messages.append(dict(role="user", content=user_feedback))
-
-    kwargs = dict(
-        model="gpt-4o",
-        n=1,
-        messages=messages,
-    )
-    logger.debug(
-        f"Getting openai output using params: \n{json.dumps(kwargs, indent=2)}"
-    )
-
-    cmp = client.chat.completions.create(**kwargs)  # type: ignore
+    kwargs = dict(model="gpt-4o", n=1)
+    logger.debug(f"Using params for OpenAI: \n{json.dumps(kwargs, indent=2)}")
+    cmp = client.chat.completions.create(messages=messages, **kwargs)  # type: ignore
     text = cmp.choices[0].message.content
     assert text is not None
-
     return text
 
 
 def assemble_letter(
-    text: str,
-    candidate_name: str,
-    company: str | None,
+    letter_config: LetterConfig,
+    letter_text: str,
     heading_text: str | None,
-    sig_path: Path | None,
 ) -> str:
+    logger.debug("Assembling letter")
     parts = []
 
     if heading_text:
         parts.append(heading_text)
 
-    if company:
-        parts.append(f"To the Hiring Team at {company}:")
+    if letter_config.company:
+        parts.append(f"To the Hiring Team at {letter_config.company}:")
     else:
         parts.append("Dear Hiring Manager,")
 
     parts.append("")
-    parts.append(text)
+    parts.append(letter_text)
     parts.append("")
     parts.append("Best regards,")
     parts.append("")
 
-    if sig_path:
-        parts.append(f"![signature](file://{sig_path.absolute()})")
+    if letter_config.sig_path:
+        parts.append(f"![signature](file://{letter_config.sig_path.absolute()})")
         parts.append("")
 
-    parts.append(candidate_name)
+    parts.append(letter_config.candidate_name)
 
     return "\n".join(parts)
 
 
-def edit_letter(text: str) -> str:
-    logger.debug("Editing generated letter")
-    with tempfile.NamedTemporaryFile() as tmp_file:
-        tmp_path = Path(tmp_file.name)
-        tmp_path.write_text(text)
-        subprocess.run([os.environ["EDITOR"], str(tmp_path)])
-        return tmp_path.read_text()
+def show_letter(letter_text: str, footer_text: str | None = None):
+    terminal_message(letter_text, subject="Generated Letter", markdown=True, footer=footer_text)
 
 
-def asset_path(file_name: str):
-    return resources.files(f"{__package__}.assets") / file_name
+def request_feedback(letter_text: str) -> str:
+    show_letter(letter_text)
+    accepted = typer.confirm("Are you satisfied with the letter?", default=True)
+    if accepted:
+        return ""
+    return typer.prompt("What can I do to fix it?", default="Just try again")
 
 
-def write_letter(
-    text: str,
-    filename_prefix: str,
-    company: str | None,
-    position: str | None,
-    dump_html: bool,
-) -> Path:
-    logger.debug("Saving letter to file")
-    name = filename_prefix
-    if company:
-        name += f"--{parameterize(company)}"
-    if position:
-        name += f"--{parameterize(position)}"
-    pdf_path = Path(f"{name}.pdf")
+def should_render(cli_param: bool | None) -> bool:
+    if cli_param is not None:
+        return cli_param
+    return typer.confirm("Should I render the letter to PDF?", default=True)
 
-    html_content = markdown(text)
 
-    if dump_html:
-        html_path = Path(f"{name}.html")
-        logger.debug(f"Dumping html to {html_path}")
-        html_path.write_text(html_content)
+def read_cached_letter(letter_config: LetterConfig) -> str:
+    return letter_config.cache_path.read_text()
 
-    css_paths = [asset_path("styles.css")]
-    html = HTML(string=html_content)
-    html.write_pdf(pdf_path, stylesheets=css_paths)
-    return pdf_path
+
+def write_cached_letter(letter_config: LetterConfig, letter_text):
+    logger.debug(f"Writing current version of letter to {letter_config.cache_path}")
+    letter_config.cache_path.write_text(letter_text)
+
+
+def edit_letter(letter_config: LetterConfig, letter_text: str):
+    edit = typer.confirm("Would you like to edit the letter?", default=False)
+    if not edit:
+        return letter_text
+
+    write_cached_letter(letter_config, letter_text)
+    logger.debug(f"Editing generated letter at {letter_config.cache_path}")
+    subprocess.run([letter_config.editor_command, str(letter_config.cache_path)])
+    return read_cached_letter(letter_config)
+
+
+def get_file_stem(letter_config: LetterConfig) -> str:
+    stem = letter_config.filename_prefix
+    if letter_config.company:
+        stem += f"--{parameterize(letter_config.company)}"
+    if letter_config.position:
+        stem += f"--{parameterize(letter_config.position)}"
+    return stem
+
+
+def build_letter(prompt_config: PromptConfig, letter_config: LetterConfig) -> str:
+    logger.debug("Building letter")
+    full_text = ""
+    heading_text: str | None = pull_heading(letter_config)
+    if letter_config.fake:
+        logger.debug("Using fake text for letter")
+        letter_text = asset_path("fake.txt").read_text()
+        full_text = assemble_letter(letter_config, letter_text, heading_text)
+        write_cached_letter(letter_config, full_text)
+    else:
+        logger.debug("Generating letter")
+        posting_text: str = pull_posting(letter_config)
+        resume_text: str = pull_resume(letter_config)
+        example_text: str | None = pull_example(letter_config)
+
+        accepted = False
+        reprompts: list[Reprompt] = []
+
+        while not accepted:
+            letter_text = generate_letter(
+                prompt_config, letter_config, reprompts, posting_text, resume_text, example_text
+            )
+            full_text = assemble_letter(letter_config, letter_text, heading_text)
+            write_cached_letter(letter_config, full_text)
+            feedback = request_feedback(full_text)
+            if feedback:
+                logger.debug("Letter was not accepted, requesting feedback")
+                reprompts.append(Reprompt(old_letter=full_text, user_feedback=feedback))
+            else:
+                accepted = True
+
+    full_text = edit_letter(letter_config, full_text)
+
+    return full_text
 
 
 cli = typer.Typer()
@@ -212,100 +203,71 @@ cli = typer.Typer()
 
 @cli.callback(invoke_without_command=True)
 @handle_abort
+@init_cache
 @attach_settings
 def generate(
     ctx: typer.Context,
     posting_url: Annotated[str, typer.Argument(help="The URL of the job posting.")],
-    company: Annotated[
-        str | None, typer.Option(help="The name of the company.")
-    ] = None,
+    company: Annotated[str | None, typer.Option(help="The name of the company.")] = None,
     position: Annotated[str | None, typer.Option(help="The title for the job.")] = None,
-    example_letter: Annotated[
-        Path | None, typer.Option(help="An example letter to use as a reference.")
-    ] = None,
-    dump_html: Annotated[
-        bool, typer.Option(help="[FOR DEBUGGING] Dump generated HTML to a file.")
-    ] = False,
+    example_letter: Annotated[Path | None, typer.Option(help="An example letter to use as a reference.")] = None,
     fake: Annotated[
         bool,
-        typer.Option(
-            help="[FOR DEBUGGING] Use a fake letter body instead of fetching one from OpenAI."
-        ),
+        typer.Option(help="[FOR DEBUGGING] Use a fake letter body instead of fetching one from OpenAI."),
     ] = False,
+    render: Annotated[
+        bool | None,
+        typer.Option(
+            help="""
+                If this option is provided, render (or don't render) the generated Markdown to PDF.
+                If not provided, you will be asked if you want to render it after the letter is generated.
+            """,
+        ),
+    ] = None,
 ):
-    settings = ctx.obj.settings
-    resume_path: Path = settings.resume_path
-    candidate_name = settings.candidate_name
-    filename_prefix = settings.filename_prefix
-    heading_path = settings.heading_path
-    sig_path = settings.sig_path
+    """
+    Generate a cover letter for a job posting listed at the provided URL.
 
-    logger.debug(f"Loading resume from {resume_path}")
-    resume_text = resume_path.read_text()
-
-    heading_text: str | None = None
-    if heading_path:
-        logger.debug(f"Loading heading text from {heading_path}")
-        heading_text = heading_path.read_text()
-
-    example_text: str | None = None
-    if example_letter:
-        logger.debug(f"Loading example text from {example_letter}")
-        example_text = example_letter.read_text()
-
-    logger.debug(f"Pulling posting from {posting_url}")
-    response = httpx.get(posting_url)
-    response.raise_for_status()
-    posting_text = response.text
-
-    logger.debug("Generating letter")
-
-    api_key = ctx.obj.settings.openai_api_key
-
-    generate_args: list[Any] = [api_key, posting_text, resume_text]
-    generate_kwargs: dict[str, Any] = dict(
-        example_text=example_text,
+    This command will use an LLM prompt (either builtin or user provided in config) to generate
+    a cover letter (in Markdown). You may also render the letter into a styled PDF.
+    """
+    prompt_config = PromptConfig(
+        dev_prompt_path=ctx.obj.settings.dev_prompt_path,
+        user_prompt_template_path=ctx.obj.settings.user_prompt_template_path,
     )
 
-    assemble_args = [
-        candidate_name,
-        company,
-        heading_text,
-        sig_path,
-    ]
+    letter_config = LetterConfig(
+        resume_path=ctx.obj.settings.resume_path,
+        candidate_name=ctx.obj.settings.candidate_name,
+        filename_prefix=ctx.obj.settings.filename_prefix,
+        openai_api_key=ctx.obj.settings.openai_api_key,
+        cache_path=BACKUP_DIR.joinpath(f".{ctx.obj.timestamp}.md"),
+        editor_command=ctx.obj.settings.editor_command,
+        sig_path=ctx.obj.settings.sig_path,
+        heading_path=ctx.obj.settings.heading_path,
+        example_path=example_letter,
+        output_directory=ctx.obj.settings.output_directory,
+        company=company,
+        position=position,
+        posting_url=posting_url,
+        fake=fake,
+    )
 
-    if fake:
-        text = asset_path("fake.txt").read_text()
+    letter_text: str = build_letter(prompt_config, letter_config)
+
+    if should_render(render):
+        render_config = RenderConfig(
+            file_stem=get_file_stem(letter_config),
+            timestamp=ctx.obj.timestamp,
+            output_directory=ctx.obj.settings.output_directory,
+        )
+        pdf_path = render_letter(letter_text, render_config)
+        simple_message(f"Letter saved to {pdf_path}")
     else:
-        text = generate_letter(*generate_args, **generate_kwargs)
-
-    full_text = assemble_letter(text, *assemble_args)
-
-    accepted = False
-    reprompts = []
-
-    while not accepted:
-        terminal_message(full_text, subject="Generated Letter", markdown=True)
-
-        accepted = typer.confirm("Are you satisfied with the letter?", default=True)
-
-        if not accepted:
-            reprompts.append(
-                (
-                    text,
-                    typer.prompt("What can I do to fix it?", default="Just try again"),
-                )
-            )
-            generate_kwargs["reprompts"] = reprompts
-
-            logger.debug("Regenerating letter based on feedback")
-            text = generate_letter(*generate_args, **generate_kwargs)
-            full_text = assemble_letter(text, *assemble_args)
-
-    edit = typer.confirm("Would you like to edit the letter?", default=False)
-    if edit:
-        edit_letter(full_text)
-
-    pdf_path = write_letter(full_text, filename_prefix, company, position, dump_html)
-
-    simple_message(f"Cover letter saved to {pdf_path}")
+        show_kwargs = {}
+        try:
+            pyperclip.copy(letter_text)
+            show_kwargs["footer_text"] = "Letter copied to clipboard"
+        except Exception as exc:
+            logger.debug(f"Could not copy letter to clipboard: {exc}")
+        show_letter(letter_text, **show_kwargs)