+++ src/smart_letters/main.py
@@ -5,6 +5,7 @@ from smart_letters.config import cli as config_cli
 from smart_letters.format import terminal_message
 from smart_letters.generate import cli as generate_cli
 from smart_letters.logging import init_logs
+from smart_letters.render import cli as render_cli
 from smart_letters.schemas import CliContext
 from smart_letters.version import show_version
 
@@ -46,6 +47,7 @@ def main(
 
 cli.add_typer(config_cli, name="config")
 cli.add_typer(generate_cli, name="generate")
+cli.add_typer(render_cli, name="render")
 
 
 if __name__ == "__main__":